Refactor: Extract common traits to core-traits crate [DRAFT]
This commit introduces a new `core-traits` crate and moves shared traits
(`ExecutionService`, `Metastore`, `WorksheetsStore`) and their associated
data structures and error types into it from `core-executor`,
`core-metastore`, and `core-history` respectively.

This is part of an effort to decouple crates and improve API boundaries
as outlined in issue #865.

Note: This is a DRAFT submission. Compilation was skipped
as per your request. Earlier compilation checks revealed issues with
Rust edition settings, Cargo feature propagation, dependency definitions,
and a Rust compiler version mismatch required by the `datafusion-functions`
dependency. These will need to be addressed.
