+++ crates/core-history/src/entities/mod.rs
@@ -1,7 +1,7 @@
-pub mod query;
-pub mod worksheet;
-pub mod worksheet_query_ref;
+// pub mod query; // Removed as query.rs was deleted
+// pub mod worksheet; // Removed as worksheet.rs was deleted
+pub mod worksheet_query_ref; // This module remains as is.
 
-pub use query::*;
-pub use worksheet::*;
+// pub use query::*; // Removed
+// pub use worksheet::*; // Removed
 pub use worksheet_query_ref::*;
